% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_probe_regions.R
\name{extract.probe.regions}
\alias{extract.probe.regions}
\title{test documentation for extract.probe.regions}
\usage{
extract.probe.regions(manifest_anno_object)
}
\description{
test documentation for extract.probe.regions
}
