\name{manifest_anno_object}
\docType{class}
\alias{manifest_anno_object}
\title{manifest_anno_object objects}
\usage{
  manifest_anno_object
}
\description{
manifest_anno_object is an example manifest object that has been properly transformed. 
The oject must have chr prefixes in the chromosome column, probes with no position information have been removed, strand designations are + and -, probe types are I and II, and column names must include CpG_id, chr, CpG_pos, strand, type.
}
