\name{SNP_bed}
\alias{SNP_bed}
\title{SNP_bed}
\usage{
  data(SNP_bed)

}
\description{
SNP_bed is an example data frame containing BED-formatted information for an exmaple subset of 1048 SNPs with Fst > .3 between African and European super-populations, as ascertained in the 1000 Genomes panel https://www.internationalgenome.org/category/phase-3/. 
}
